+++ codex/serializers/metadata.py
@@ -1,5 +1,5 @@
 """Codex Serializers for the metadata box."""
-from rest_framework.serializers import CharField
+from rest_framework.serializers import CharField, IntegerField
 
 from codex.serializers.mixins import (
     BrowserAggregateSerializerMixin,
@@ -16,3 +16,4 @@ class MetadataSerializer(BrowserAggregateSerializerMixin, ComicSerializer):
     """Aggregate stats for the comics selected in the metadata dialog."""
 
     filename = CharField(read_only=True)
+    parent_folder_pk = IntegerField(read_only=True, required=False)