+++ codex/views/browser/metadata.py
@@ -50,20 +50,13 @@ class MetadataView(BrowserAnnotationsView):
         "created_at",
     }
     _COMIC_VALUE_FIELDS_CONFLICTING_PREFIX = "conflict_"
-    _COMIC_FK_FIELDS = {
-        "publisher",
-        "imprint",
-        "series",
-        "volume",
-    }
     _COMIC_FK_FIELDS_MAP = {
-        "f": _COMIC_FK_FIELDS,
-        "p": _COMIC_FK_FIELDS.difference(frozenset(["publisher"])),
-        "i": _COMIC_FK_FIELDS.difference(frozenset(["imprint"])),
-        "s": _COMIC_FK_FIELDS.difference(frozenset(["series"])),
-        "v": _COMIC_FK_FIELDS.difference(frozenset(["volume"])),
-        "c": _COMIC_FK_FIELDS,
+        "p": "publisher",
+        "i": "imprint",
+        "s": "series",
+        "v": "volume",
     }
+    _COMIC_FK_FIELDS = frozenset(_COMIC_FK_FIELDS_MAP.values())
     _COMIC_FK_ANNOTATION_PREFIX = "fk_"
     _COMIC_RELATED_VALUE_FIELDS = {"series__volume_count", "volume__issue_count"}
     _COUNT_FIELD_MAP = {
@@ -127,16 +120,17 @@ class MetadataView(BrowserAnnotationsView):
         count_field, count_src_field_name = op_field_names
         val[count_field] = obj[count_src_field_name]
 
-    @classmethod
-    def _field_copy(cls, obj, fields, prefix, fk_field=None):
+    def _field_copy(self, obj, fields, prefix, fk_field=None):
         """Copy annotation fields into comic type fields."""
         for field_name in fields:
             src_field_name = f"{prefix}{field_name}"
             val = obj[src_field_name]
             if fk_field:
                 val = {fk_field: val}
+                if pk := obj.get(field_name + "_id"):
+                    val["pk"] = pk
                 # Add special count fields
-                cls._field_copy_fk_count(field_name, obj, val)
+                self._field_copy_fk_count(field_name, obj, val)
             obj[field_name] = val
 
     def _annotate_aggregates(self, qs):
@@ -186,6 +180,7 @@ class MetadataView(BrowserAnnotationsView):
         """Annotate for the filename function."""
         if not self.is_model_comic:
             return qs
+        qs = qs.annotate(parent_folder_pk=F("parent_folder_id"))
         return qs.annotate(series_name=F("series__name"), volume_name=F("volume__name"))
 
     def _query_m2m_intersections(self, simple_qs):
@@ -225,7 +220,7 @@ class MetadataView(BrowserAnnotationsView):
             reason = f"Cannot get metadata for {self.group=}"
             raise ValueError(reason)
         group_field = self.model.__name__.lower()
-        obj[group_field] = {"name": obj["name"]}
+        obj[group_field] = {"pk": obj["id"], "name": obj["name"]}
         comic_fk_fields = copy(self._COMIC_FK_FIELDS_MAP[self.group])
         self._field_copy(
             obj,